auto (new_habit.c) 450 3825

+++ sql/base.sql
@@ -1,6 +1,6 @@
 create table if not exists habits (
     id INTEGER primary key,
-    importance integer NOT NULL,
+    importance REAL NOT NULL,
     description text NOT NULL,
     enabled boolean NOT NULL DEFAULT 1,
     frequency INTEGER NOT NULL, -- in milliseconds
